requirements.txt
django
django-rest-framework
psycopg2-binary
django-oauth-toolkit
djangorestframework_simplejwt
Pillow
